8246112: Remove build-time and run-time checks for clock_gettime and CLOCK_MONOTONIC
