+++ src/hotspot/os/bsd/os_bsd.inline.hpp
@@ -109,14 +109,6 @@ inline struct hostent* os::get_host_by_name(char* name) {
   return ::gethostbyname(name);
 }
 
-inline bool os::supports_monotonic_clock() {
-#ifdef __APPLE__
-  return true;
-#else
-  return Bsd::_clock_gettime != NULL;
-#endif
-}
-
 inline void os::exit(int num) {
   ::exit(num);
 }
